pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
psycopg2-binary = "^2.9.6"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
celery = "^5.2.7"
redis = "^4.5.5"
pyyaml = "^6.0"
uritemplate = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
